Create index.html
This page is the main page for the application. The file has been tested on my local and works. We can add images once we get data solidified as the images need to be on the server (see me for questions about this if you have them) in order to be viewable from every machine. Please see email.