Create InformationWebpages.html
This page is the subsequent data pages for the application. We will use this as a template for the other pages within our application. The file has been tested on my local and works. We can add images once we get data solidified as the images need to be on the server (see me for questions about this if you have them) in order to be viewable from every machine.